backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# LlamaIndex
llama-index==0.9.14
llama-index-llms-gemini==0.1.3
llama-index-embeddings-huggingface==0.1.4

# ML/NLP
torch==2.1.0
sentence-transformers==2.2.2
transformers==4.35.2